generic download menu component for easier reading
